requirements.txt
cmu-graphics==1.1.47
cmu-graphics-helpers==0.1.3
cppyy==3.5.0
cppyy-backend==1.15.3
cppyy-cling==6.32.8
CPyCppyy==1.13.0
numpy==2.4.3
pillow==12.2.0
pycairo==1.29.0
pygame==2.6.1
pygame-ce==2.5.7
pygame_widgets==1.3.2
pyperclip==1.11.0
